# apps/go_backend/docs/swagger/swagger.yaml
basePath: /
definitions:
  auth.AuthResponse:
    description: Authentication response with JWT token
    properties:
      is_admin:
        example: false
        type: boolean
      token:
        example: eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        example: user:abc123
        type: string
    type: object
  auth.LoginRequest:
    description: Login request payload
    properties:
      password:
        example: adminadmin
        minLength: 6
        type: string
      username:
        example: admin@example.com
        minLength: 1
        type: string
    required:
    - password
    - username
    type: object
  auth.RegisterRequest:
    description: Registration request payload
    properties:
      password:
        example: password123
        minLength: 6
        type: string
      username:
        example: user@example.com
        type: string
    required:
    - password
    - username
    type: object
  categories.Category:
    properties:
      color:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  categories.CategoryPageResponse:
    properties:
      has_more:
        type: boolean
      items:
        items:
          $ref: '#/definitions/categories.Category'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  categories.CreateRequest:
    description: Request payload for creating a category
    properties:
      color:
        example: '#3B82F6'
        maxLength: 50
        minLength: 1
        type: string
      name:
        example: Work
        maxLength: 100
        minLength: 1
        type: string
    required:
    - color
    - name
    type: object
  categories.UpdateRequest:
    description: Request payload for updating a category
    properties:
      color:
        example: '#10B981'
        maxLength: 50
        minLength: 1
        type: string
      name:
        example: Personal
        maxLength: 100
        minLength: 1
        type: string
    type: object
  errors.ErrorCode:
    enum:
    - BAD_REQUEST
    - VALIDATION_FAILED
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - CONFLICT
    - TOO_MANY_REQUESTS
    - INTERNAL_ERROR
    - DATABASE_ERROR
    - SERVICE_UNAVAILABLE
    type: string
    x-enum-varnames:
    - CodeBadRequest
    - CodeValidationFailed
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeConflict
    - CodeTooManyRequests
    - CodeInternalError
    - CodeDatabaseError
    - CodeServiceError
...
  response.APIError:
    properties:
      code:
        $ref: '#/definitions/errors.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.APIError'
    type: object
  response.OperationMessage:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: Daily journal backend powered by SurrealDB.
  title: Lucid Logs API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: User login
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user and return JWT token
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: User registration
      tags:
      - auth
  /api/v1/categories:
    get:
      description: Get paginated categories for the authenticated user
      parameters:
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Items to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/categories.CategoryPageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a category for organizing tasks
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/categories.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/categories.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - categories
  /api/v1/categories/{id}:
    delete:
      description: Soft delete a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OperationMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - categories
    get:
      description: Get a single category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/categories.Category'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get category
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update the name or color of a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/categories.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/categories.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - categories
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"